requirements.txt
# PPT 生成依赖
python-pptx>=0.6.21

# 视频生成依赖
edge-tts>=6.1.0
Pillow>=10.0.0

# PDF 中转方案依赖（可选但推荐）
reportlab>=4.0.0

# Word 文档生成依赖
python-docx>=0.8.11

# 大模型接口依赖（项目依赖）
# 需要确保 llm_services/qwen_engine.py 模块可用